refactor(columnmanager): separate static functions into new file
