+++ src/os/ui/column/columnmanager.js
@@ -462,68 +462,3 @@ os.ui.column.ColumnManagerCtrl.prototype.handleKeyEvent_ = function(event) {
     }
   }
 };
-
-
-/**
- * Launches a column manager window with the given columns
- *
- * @param {Array.<os.data.ColumnDefinition>} columns
- * @param {Function} callback
- */
-os.ui.column.launchColumnManager = function(columns, callback) {
-  var scopeOptions = {
-    'columns': columns,
-    'acceptCallback': callback
-  };
-
-  var windowOptions = {
-    'label': 'Column Manager',
-    'icon': 'fa fa-columns',
-    'x': 'center',
-    'y': 'center',
-    'width': '600',
-    'min-width': '500',
-    'max-width': '700',
-    'height': '400',
-    'min-height': '350',
-    'max-height': '1000',
-    'show-close': true,
-    'modal': true
-  };
-
-  var template = '<column-manager columns="columns" accept-callback="acceptCallback" ></column-manager>';
-  os.ui.window.create(windowOptions, template, undefined, undefined, undefined, scopeOptions);
-};
-
-
-/**
- * Launches a column manager window with the given columns
- *
- * @param {Array.<os.data.ColumnDefinition>} columns
- * @param {string} header Dialog header
- * @param {Function} callback
- */
-os.ui.column.launchColumnManagerWithShownCallback = function(columns, header, callback) {
-  var scopeOptions = {
-    'columns': columns,
-    'shownCallback': callback
-  };
-
-  var windowOptions = {
-    'label': header,
-    'icon': 'fa fa-columns',
-    'x': 'center',
-    'y': 'center',
-    'width': '600',
-    'min-width': '500',
-    'max-width': '700',
-    'height': '400',
-    'min-height': '350',
-    'max-height': '1000',
-    'show-close': true,
-    'modal': true
-  };
-
-  var template = '<column-manager columns="columns" shown-callback="shownCallback"></column-manager>';
-  os.ui.window.create(windowOptions, template, undefined, undefined, undefined, scopeOptions);
-};
